[Refactor] #963 Separated take_turn() and free_turn() from player-status.cpp/h to player-energy.cpp/h
