+++ src/action/travel-execution.cpp
@@ -13,8 +13,8 @@
 #include "game-option/special-options.h"
 #include "grid/feature.h"
 #include "grid/grid.h"
+#include "player-status/player-energy.h"
 #include "player/player-move.h"
-#include "player/player-status.h"
 #include "system/floor-type-definition.h"
 #include "system/monster-type-definition.h"
 #include "system/player-type-definition.h"
