+++ src/action/tunnel-execution.cpp
@@ -11,8 +11,8 @@
 #include "main/sound-definitions-table.h"
 #include "main/sound-of-music.h"
 #include "player-info/avatar.h"
+#include "player-status/player-energy.h"
 #include "player/player-move.h"
-#include "player/player-status.h"
 #include "system/floor-type-definition.h"
 #include "system/player-type-definition.h"
 #include "util/bit-flags-calculator.h"
