+++ src/cmd-action/cmd-open-close.cpp
@@ -12,8 +12,8 @@
 #include "inventory/inventory-object.h"
 #include "inventory/inventory-slot-types.h"
 #include "io/input-key-requester.h"
+#include "player-status/player-energy.h"
 #include "player/attack-defense-types.h"
-#include "player/player-status.h"
 #include "player/special-defense-types.h"
 #include "specific-object/chest.h"
 #include "status/action-setter.h"
