+++ src/cmd-action/cmd-spell.cpp
@@ -32,6 +32,7 @@
 #include "object/item-use-flags.h"
 #include "player-info/avatar.h"
 #include "player-info/self-info.h"
+#include "player-status/player-energy.h"
 #include "player/attack-defense-types.h"
 #include "player/eldritch-horror.h"
 #include "player/player-class.h"
