+++ src/cmd-action/cmd-tunnel.cpp
@@ -7,8 +7,8 @@
 #include "grid/feature.h"
 #include "grid/grid.h"
 #include "io/input-key-requester.h"
+#include "player-status/player-energy.h"
 #include "player/attack-defense-types.h"
-#include "player/player-status.h"
 #include "player/special-defense-types.h"
 #include "status/action-setter.h"
 #include "system/floor-type-definition.h"
