+++ src/cmd-building/cmd-building.cpp
@@ -55,8 +55,8 @@
 #include "object-hook/hook-weapon.h"
 #include "object/item-tester-hooker.h"
 #include "player-info/avatar.h"
+#include "player-status/player-energy.h"
 #include "player/player-personality-types.h"
-#include "player/player-status.h"
 #include "spell-kind/spells-perception.h"
 #include "spell-kind/spells-world.h"
 #include "spell/spells-status.h"
