+++ src/cmd-item/cmd-destroy.cpp
@@ -21,8 +21,8 @@
 #include "object/object-stack.h"
 #include "object/object-value.h"
 #include "player-info/avatar.h"
+#include "player-status/player-energy.h"
 #include "player/attack-defense-types.h"
-#include "player/player-status.h"
 #include "player/special-defense-types.h"
 #include "racial/racial-android.h"
 #include "realm/realm-names-table.h"
