+++ src/cmd-item/cmd-eat.cpp
@@ -24,6 +24,7 @@
 #include "object/object-kind-hook.h"
 #include "object/object-kind.h"
 #include "perception/object-perception.h"
+#include "player-status/player-energy.h"
 #include "player-info/avatar.h"
 #include "player/attack-defense-types.h"
 #include "player/digestion-processor.h"
@@ -32,7 +33,6 @@
 #include "player/player-damage.h"
 #include "player/player-race-types.h"
 #include "player/player-status-flags.h"
-#include "player/player-status.h"
 #include "player/special-defense-types.h"
 #include "spell-realm/spells-hex.h"
 #include "spell-realm/spells-song.h"
