+++ src/cmd-item/cmd-magiceat.cpp
@@ -68,9 +68,9 @@
 #include "object/object-kind.h"
 #include "sv-definition/sv-other-types.h"
 #include "sv-definition/sv-rod-types.h"
+#include "player-status/player-energy.h"
 #include "player-info/avatar.h"
 #include "player/player-class.h"
-#include "player/player-status.h"
 #include "player/player-status-table.h"
 #include "spell/spell-info.h"
 #include "system/player-type-definition.h"
