+++ src/cmd-item/cmd-usestaff.cpp
@@ -18,10 +18,10 @@
 #include "perception/object-perception.h"
 #include "player/attack-defense-types.h"
 #include "player-info/avatar.h"
+#include "player-status/player-energy.h"
 #include "player/player-class.h"
 #include "player/player-race-types.h"
 #include "player/player-race.h"
-#include "player/player-status.h"
 #include "player/player-status-flags.h"
 #include "player/special-defense-types.h"
 #include "spell-kind/earthquake.h"
