+++ src/cmd-item/cmd-zaprod.cpp
@@ -13,8 +13,8 @@
 #include "perception/object-perception.h"
 #include "player/attack-defense-types.h"
 #include "player-info/avatar.h"
+#include "player-status/player-energy.h"
 #include "player/player-class.h"
-#include "player/player-status.h"
 #include "player/special-defense-types.h"
 #include "spell-kind/spells-beam.h"
 #include "spell-kind/spells-detection.h"
