+++ src/dungeon/quest.cpp
@@ -26,6 +26,7 @@
 #include "object-enchant/item-apply-magic.h"
 #include "object-enchant/trg-types.h"
 #include "object/object-generator.h"
+#include "player-status/player-energy.h"
 #include "player/player-personality-types.h"
 #include "player/player-status.h"
 #include "system/artifact-type-definition.h"
