+++ src/mind/mind-blue-mage.cpp
@@ -12,8 +12,8 @@
 #include "monster-race/race-ability-mask.h"
 #include "mspell/monster-power-table.h"
 #include "player-info/avatar.h"
+#include "player-status/player-energy.h"
 #include "player/player-status-table.h"
-#include "player/player-status.h"
 #include "realm/realm-types.h"
 #include "spell/spell-info.h"
 #include "status/bad-status-setter.h"
