+++ src/object-use/quaff-execution.cpp
@@ -20,15 +20,15 @@
 #include "object/object-info.h"
 #include "object/object-kind.h"
 #include "perception/object-perception.h"
-#include "player-info/self-info.h"
-#include "player/attack-defense-types.h"
 #include "player-info/avatar.h"
-#include "player/player-status-flags.h"
+#include "player-info/self-info.h"
+#include "player-status/player-energy.h"
+#include "player/attack-defense-types.h"
 #include "player/digestion-processor.h"
 #include "player/eldritch-horror.h"
 #include "player/mimic-info-table.h"
 #include "player/player-damage.h"
-#include "player/player-status.h"
+#include "player/player-status-flags.h"
 #include "realm/realm-hex-numbers.h"
 #include "spell-kind/spells-detection.h"
 #include "spell-kind/spells-floor.h"
