+++ src/spell/spells-status.cpp
@@ -28,9 +28,9 @@
 #include "object/object-kind-hook.h"
 #include "object/object-kind.h"
 #include "player-info/avatar.h"
+#include "player-status/player-energy.h"
 #include "player/attack-defense-types.h"
 #include "player/player-class.h"
-#include "player/player-status.h"
 #include "spell-kind/spells-launcher.h"
 #include "spell-kind/spells-teleport.h"
 #include "spell-kind/spells-world.h"
