+++ src/status/action-setter.cpp
@@ -14,8 +14,8 @@
 #include "status/action-setter.h"
 #include "core/player-redraw-types.h"
 #include "core/player-update-types.h"
+#include "player-status/player-energy.h"
 #include "player/attack-defense-types.h"
-#include "player/player-status.h"
 #include "player/special-defense-types.h"
 #include "spell-realm/spells-hex.h"
 #include "spell-realm/spells-song.h"
