+++ src/store/cmd-store.cpp
@@ -20,7 +20,7 @@
 #include "main/sound-of-music.h"
 #include "object/object-generator.h"
 #include "object/object-info.h"
-#include "player/player-status.h"
+#include "player-status/player-energy.h"
 #include "store/cmd-store.h"
 #include "store/home.h"
 #include "store/store-key-processor.h"
