+++ src/world/world-turn-processor.cpp
@@ -27,8 +27,8 @@
 #include "mutation/mutation-processor.h"
 #include "object/lite-processor.h"
 #include "perception/simple-perception.h"
+#include "player-status/player-energy.h"
 #include "player/digestion-processor.h"
-#include "player/player-status.h"
 #include "store/store-owners.h"
 #include "store/store-util.h"
 #include "store/store.h"
